+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.0 (in development)
+
+- Added `GridSearch` and moved properties `search`, `searchParamName`, `searchInputOptions` and `searchUrl` as well as
+  methods `getSearchInput`, and `getSearchKeywords` to the new `GridView::$search` property
+
 ## 2.4.2 (Dec 17, 2024)
 
 - Changed `Skipping unsafe attribute` log level to `debug`
@@ -54,7 +59,7 @@
 ## 2.3.15 (Sep 17, 2024)
 
 - Added the option to remove empty span `tags` from HtmlValidator output, defaults to false
-- Enhanced `MigrationController` to not offer to create a new user if an error occured during migration
+- Enhanced `MigrationController` to not offer to create a new user if an error occurred during migration
 - Fixed translations in `AuthItemGridView` for message sources with `forceTranslation` set to `true`
 
 ## 2.3.14 (Sep 11, 2024)